refactor(Card):extract rating into Rating component
Extract `Array.createRange` into `createRange` util

[Continues #3]
